Day 12: modify POST /tasks, GET /tasks, DELETE /tasks

+++ logic.py
@@ -13,7 +13,7 @@ def tasks_row_list(rows):
 
 
 def tasks_row_dict(row):
-	return {"id": row["id"], "title": row["title"], "done": bool(row["done"])}
+	return {"id": row["id"], "title": row["title"], "done": bool(row["done"]), "user_id": row["user_id"]}
 	
 	
 def users_row_list(rows):
@@ -29,7 +29,7 @@ def find_task(task_id):
 	set_row_factory(conn)
 	cur = conn.cursor()
 
-	cur.execute("SELECT id, title, done FROM tasks WHERE id = :id", {"id": task_id})
+	cur.execute("SELECT id, title, done, user_id FROM tasks WHERE id = :id", {"id": task_id})
 	
 	task =  cur.fetchone()
 	
@@ -147,45 +147,46 @@ def delete_user(user_id):
 	return user
 	
 # Tasks area
-def add_task(title):
+def add_task(user_id, title):
 	conn = get_connection()
 	set_row_factory(conn)
 	cur = conn.cursor()
 	
-	cur.execute("INSERT INTO tasks (title, done) VALUES (:title, :done)", {"title": title, "done": False})
+	cur.execute("INSERT INTO tasks (title, done, user_id) VALUES (:title, :done, :user_id)", {"title": title, "done": False, "user_id": user_id})
 	conn.commit()
 	
-	cur.execute("SELECT id, title, done FROM tasks WHERE id = :id", {"id": cur.lastrowid})
+	cur.execute("SELECT id, title, done, user_id FROM tasks WHERE id = :id", {"id": cur.lastrowid})
 	task = cur.fetchone()
 	conn.close()
 	
-	return task
+	return tasks_row_dict(task)
 
 
-def get_tasks(done=None, search=None, sort=None, page=1, limit=10):
+def get_tasks(user_id, done=None, search=None, sort=None, page=1, limit=10):
 	conn = get_connection()
 	set_row_factory(conn)
 	cur = conn.cursor()
 	
 	offset = (page - 1) * limit
 	
-	query = "SELECT id, title, done FROM tasks "
+	query = "SELECT id, title, done, user_id FROM tasks "
 	params = {}
 	conditions = []
 	
+	if user_id is not None:
+		conditions.append("user_id = :user_id")
+		params["user_id"] = user_id
 	if done is not None:
 		conditions.append("done = :done")
 		params["done"] = int(done)
-	#return done
 	if search is not None:
 		conditions.append("title LIKE :search")
 		params["search"] = f"%{search}%"
-		
+	
 	if conditions:
 		query += " WHERE " + " AND ".join(conditions)
 		
 	if sort is not None:
-		#return sort
 		query += f" ORDER BY id {sort} "
 		#params["sort"] = f"{sort}"
 	
@@ -215,14 +216,14 @@ def update_task(task_id, title, done):
 		cur.execute("UPDATE tasks SET title = :title WHERE id = :id", {"title": title, "id": task_id})
 	if done is not None:
 		cur.execute("UPDATE tasks SET done = :done WHERE id = :id", {"done": done, "id": task_id})
-		
+	
 	conn.commit()
 	conn.close()
 	
 	return find_task(task_id)
 
 	
-def delete_task(task_id):
+def delete_task(user_id, task_id):
 	conn = get_connection()
 	set_row_factory(conn)
 	cur = conn.cursor()
@@ -231,6 +232,9 @@ def delete_task(task_id):
 	
 	if task is None:
 		return None
+		
+	if user_id != task["user_id"]:
+		return False
 		
 	cur.execute("DELETE FROM tasks WHERE id = :id", {"id": task_id})
